Resolve some comments on the PR
