+++ backend/app/mcp/mcp.py
@@ -4,8 +4,6 @@ from uuid import UUID
 from fastmcp import Context, FastMCP
 from fastmcp.server.auth import BearerAuthProvider
 from fastmcp.server.dependencies import AccessToken, get_access_token
-from fastmcp.server.middleware import Middleware, MiddlewareContext
-from posthog import Posthog
 from sqlalchemy.orm import configure_mappers
 
 from app.core.auth.auth import get_authenticated_user
@@ -27,52 +25,6 @@ from app.domains.schema_response import DomainGet
 from app.domains.service import DomainService
 from app.settings import settings
 
-# from app.core.auth.jwt import JWTToken, get_oidc
-
-
-if settings.POSTHOG_ENABLED:
-    posthog_client = Posthog(
-        project_api_key=settings.POSTHOG_API_KEY, host=settings.POSTHOG_HOST
-    )
-
-
-class LoggingMiddleware(Middleware):
-    """Middleware that logs all MCP operations."""
-
-    async def on_message(self, context: MiddlewareContext, call_next):
-        """Called for all MCP messages."""
-        entry = {
-            "timestamp": context.timestamp.isoformat(),
-            "source": context.source,
-            "type": context.type,
-            "method": context.method,
-        }
-
-        if hasattr(context.message, "__dict__"):
-            try:
-                entry["payload"] = context.message.__dict__
-            except (TypeError, ValueError):
-                entry["payload"] = "<non-serializable>"
-
-        try:
-            access_token: AccessToken = get_access_token()
-            user_id = get_authenticated_user(
-                token=JWTToken(sub="", token=f"Bearer {access_token.token}"),
-                db=next(get_db_session()),
-            ).id
-        except Exception:
-            user_id = None
-
-        if settings.POSTHOG_ENABLED:
-            # Log to PostHog
-            posthog_client.capture(
-                distinct_id=user_id,
-                event="MCP Method Call API",
-                properties=entry,
-            )
-        result = await call_next(context)
-        return result
-
 
 def initialize_models():
     """Initialize all SQLAlchemy models and resolve relationships"""
@@ -82,23 +34,25 @@ def initialize_models():
         print(f"Warning during model initialization: {e}")
 
 
-initialize_models()
+initialize_models()  # TODO Figure out if this is still needed
 
-auth = None
-if settings.OIDC_ENABLED:
-    auth = BearerAuthProvider(
-        issuer=get_oidc().authority,
-        jwks_uri=get_oidc().jwks_uri,
-    )
 
-mcp = FastMCP(name="DataProductPortalMCP", auth=auth)
+def get_auth_provider() -> Optional[BearerAuthProvider]:
+    if settings.OIDC_ENABLED:
+        return BearerAuthProvider(
+            issuer=get_oidc().authority, jwks_uri=get_oidc().jwks_uri
+        )
+    return None
+
+
+mcp = FastMCP(name="DataProductPortalMCP", auth=get_auth_provider())
 
 # ==============================================================================
 # CORE DISCOVERY & SEARCH TOOLS
 # ==============================================================================
 
 
-def get_current_user(token: str):
+def get_mcp_authenticated_user(token: str):
     user = get_authenticated_user(
         token=JWTToken(sub="", token=f"Bearer {token}"), db=next(get_db_session())
     )
@@ -112,11 +66,11 @@ def get_current_user(token: str):
 
 
 @mcp.tool
-async def get_my_data(ctx: Context) -> dict[str, Any]:
-    """Get my data from the MCP. You can use this to
+async def get_current_user(ctx: Context) -> dict[str, Any]:
+    """Get current user data from the MCP. You can use this to
     get information about the authenticated user."""
     access_token: AccessToken = get_access_token()
-    return get_current_user(token=access_token.token)
+    return get_mcp_authenticated_user(token=access_token.token)
 
 
 @mcp.tool
@@ -136,7 +90,7 @@ def universal_search(
         access_token: AccessToken = get_access_token()
 
         db = next(get_db_session())
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             results = {
                 "query": query,
@@ -306,7 +260,7 @@ def search_datasets(
     try:
         db = next(get_db_session())
         access_token: AccessToken = get_access_token()
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             # Get all datasets and filter manually
             all_datasets = DatasetService(db).get_datasets(user=user)
@@ -388,7 +342,7 @@ def get_dataset_details(dataset_id: str) -> Dict[str, Any]:
     try:
         db = next(get_db_session())
         access_token: AccessToken = get_access_token()
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             dataset = DatasetService(db).get_dataset(id=UUID(dataset_id), user=user)
 
@@ -456,7 +410,7 @@ def get_marketplace_overview() -> Dict[str, Any]:
     try:
         db = next(get_db_session())
         access_token: AccessToken = get_access_token()
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             # Get counts by querying all and taking length
             all_data_products = DataProductService(db).get_data_products()
@@ -503,7 +457,7 @@ def get_data_product_analytics(data_product_id: str) -> Dict[str, Any]:
     try:
         db = next(get_db_session())
         access_token: AccessToken = get_access_token()
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             # Get the data product using service
             data_product = DataProductService(db).get_data_product(
@@ -608,7 +562,7 @@ def get_dataset_resource(dataset_id: str) -> str:
     try:
         db = next(get_db_session())
         access_token: AccessToken = get_access_token()
-        user = get_current_user(token=access_token.token)
+        user = get_mcp_authenticated_user(token=access_token.token)
         try:
             dataset = DatasetService(db).get_dataset(id=UUID(dataset_id), user=user)
